Index files, improve performance with LRU cache

+++ source/Spinach/Trigrams/TrigramInfo.cs
@@ -2,6 +2,6 @@ namespace Spinach.Trigrams;
 
 public struct TrigramInfo
 {
-  public long Key { get; set; }
+  public int Key { get; set; }
   public long Position { get; set; }
 }
